Realized that non-forward/backward propagation doesn't make sense, as it will likely not really propagate but rather just "get stuck".
Also realized that propagating from the source region doesn't make sense.

Say you have two pixels next to each other:
	A B
A is in the source region and B is in the target region. Since A is in the source region,
it’s best NN is exactly itself. When B says “check the neighbor of the nearest neighbor
of my neighbor (A) whose offset is the same as my offset from A”, it returns B!
Clearly we don’t want to use B as B’s NN, so this doesn’t make sense to do this.